Add files via upload
This code display the Super Mart. In this code the menu card represents as a python dictionary. Initially the total cost is zero. By using if else statement it checks the condition whether the item is exist in the menu card or not which is entered by the user (buyer). On the other hand, if else ladder also checks the condition whether if the total cost is equal to or greater than 100000 so the buyer eligible for a 20% discount other wise the total cost is less than 100000 so the buyer is not eligible for a discount.